Lots More Stuff 2025_8_23
-Added Save and Load functionality
(This showed me that my door and key system was broken)
-Added lighting system
(Camera now no longer snaps to room borders to allow viewing each corridor as the character will see. There is no longer the ability to see a path from the other side of a wall)
-Added back the Doors
(Doors can be placed or removed through code. Stairs are in the works.)